+++ app/pages/pds.stories.ts
@@ -1,7 +1,7 @@
 import Pds from './pds.vue'
 import type { Meta, StoryObj } from '@storybook-vue/nuxt'
 import { pageDecorator } from '../../.storybook/decorators'
-import { pdsUsersHandler } from '../mocks/handlers'
+import { pdsUsersHandler } from '../storybook/mocks/handlers'
 
 const meta = {
   component: Pds,
